+++ tcapy_scripts/gen/dump_data_vendor_to_parquet_csv_hdf5.py
@@ -18,7 +18,7 @@ if __name__ == '__main__':
 
     from tcapy.conf.constants import Constants
 
-    data_vendor = 'dukascopy' # 'ncfx' or 'dukascopy'
+    data_vendor = 'ncfx' # 'ncfx' or 'dukascopy'
     write_large_csv = False
     write_large_hdf5_parquet = True
     return_df = True # returns the dataframe (DO NOT DO this for large datasets, as it will cause you to run out of memory)
@@ -27,7 +27,7 @@ if __name__ == '__main__':
                                 # however, when finally copying to the Arctic database, we recommend removing duplicates
                                 # otherwise, it quickly results in going out of memory
 
-    csv_folder = '/home/tcapyuser/csv_dump/'
+    csv_folder = '/home/tcapyuser/csv_dump/' + data_vendor + '/'
     constants = Constants()
 
     # Where should we dump the temporary FX data mini files and large H5 files
@@ -46,8 +46,8 @@ if __name__ == '__main__':
     temp_data_folder = '/home/tcapyuser/csv_dump/temp/'
     temp_large_data_folder = '/home/tcapyuser/csv_dump/temp/large/'
 
-    start_date_csv = '01 Jan 2016'; finish_date_csv = '07 Apr 2020'; split_size = 'monthly' # 'daily' or 'monthly'
-    # start_date_csv = '01 Jan 2004'; finish_date_csv = '01 Jan 2006';
+    start_date_csv = '01 Apr 2016'; finish_date_csv = '01 May 2020'; split_size = 'monthly' # 'daily' or 'monthly'
+    # start_date_csv = '01 Jan 2005'; finish_date_csv = '01 May 2020';
 
     if data_vendor == 'ncfx':
         from tcapy.data.databasepopulator import DatabasePopulatorNCFX as DatabasePopulator